requirements.txt
discord.py
python-dotenv
pynacl
pygifsicle
